change deployment upgrade strategy to recreate
currently we deploy using the default Rolling-update strategy.
This causes issues with the new leadership election, making it so that the old pods can't be deleted until the a new one is ready (which won't happen until both pods are up and election is won).
We are thus changing the strategy to recreate which will kill old pods and then redeploy the new ones. This has the added benefit of removing potential races between old and new election mechanisms.

Signed-off-by: Ram Lavi <ralavi@redhat.com>